pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.6",
]
